+++ test.txt
@@ -1,5 +1,5 @@
-POST / HTTP/1.1
-Host: example.com
+POST /post HTTP/1.1
+Host: localhost:8080
 Transfer-Encoding: chunked
 
 6
@@ -7,3 +7,4 @@ Hello
 6
 World!
 0
+
